Sceanrio 1
Task: Imagine you and your teammate are working on the same file, project.txt. Both of you make changes to the same lines and commit them. Let’s say your teammate’s changes are merged first. And now when you try to merge your changes, a merge conflict occurs.

Expected result: Changes from both branches should be included in the file in the main branch. The changes should be included in the sequence: first your teammate’s changes and then the changes made by you.

Steps to create a conflict:

Create a repository and add a file named project.txt.
Create two branches in the repository.
Make changes to the same line and in the same file in both branches.
Commit the changes in both branches.
Merge one branch into the main first and then try to merge the second branch.
Understand the message that appears when you try to merge the second branch and explain that message. In case of conflicts explain the steps taken to resolve the conflict.